Crate problem093

Problem 93 (Arithmetic expressions)

Solution detail

Brute-force is the easiest way to approach this problem - the question is how to implement it efficiently and ergonomically!

One approach is to consider what targets can be reached from particular sets of digits. Given a particular set of digits, all possible targets are obtained by considering all the ways of splitting that set into two non-empty subsets, taking all possible targets from the two subsets and combining them with every possible operation.

This framing of the problem leads to an iterative solution whereby the base case consists of the singleton sets containing a single digit - for them, the only possible target is the number itself.

From there, iterate over the possible sets of digits in order of increasing size, combining the answers to previous sets in all possible combinations.

Once all targets have been calculated, it's a simple matter of filtering to integer answers and doing a bit of counting.

Functions

all_ops
All results of a binary operation on the two given numbers

answer
Solve the problem, returning the answer as a String

count_consecutive
Count the consecutive integers 1, 2, ..., n that exist in the given set of targets

digit_set_to_integer
A handy conversion to represent four digits as one number

main

solve
Find the four digits which allow the most consecutive integer targets to be reached.

update
Update the target HashSet with all possible operations on two numbers from the two given HashSets.

Type Definitions

DigitSet

Target